requirements.txt
langchain 
langchain_core 
langchain_openai 
langchain_text_splitters
langchain-community
langchain-chroma
pypdf
langchain-groq
pandas
python-dotenv
streamlit